refactor(web): declare the API types by domain under api/types/
types.ts was a 1058-line wall of interfaces. The declarations move into
api/types/{providers,models,logs,stats,failover,quota,chat,settings,auth}.ts
and types.ts becomes the barrel that re-exports them, which is the shape
it already had for the shared quota types: app code keeps importing
every API type from one place, so not one of the ~100 import sites
changes.

The exported name set is identical before and after (diffed), and the
only cross-file reference the split created is models.ts importing the
three failover group-info types for DiscoveryDiff.

types.ts goes from 1058 to 30 lines and leaves the size allowlist.
