+++ scripts/ci/size-allowlist.txt
@@ -22,7 +22,6 @@ internal/proxy/model_gone.go	1163
 internal/proxy/model_gone_test.go	2739
 internal/proxy/models_test.go	2307
 internal/proxy/upstream_classify.go	804
-web/src/api/types.ts	1058
 web/src/components/__tests__/Layout.navigation.test.tsx	2558
 web/src/pages/Settings/alerts/AlertsWizard.tsx	801
 web/src/pages/Settings/alerts/steps.tsx	846
